필요없는 파일 삭제 했습니다,ai 관련 함수 이야기 했던대로 4개로 추렸습니다
